+++ vista/CHANGE-LOG.md
@@ -2,6 +2,7 @@
 
 2025-29-01 - Stefano Renzetti <stefano.renzetti3@studio.unibo.it>
 
+- Mapped `recognitionClass` and `stageClass` to dynamically asssign correct class.
 - Modified `saveRdf` in `LocalStorageAdapter.js` to drop all existing annotations to not duplicate them.
 - 👀 Implemented JSON-LD context in contexs in `LocalStorageAdapter.js` to correctly convert in N-Quads.
 - CORS error need to be fix to establish the connection with Blazegraph.